logger is in its own module now
