+++ app.js
@@ -3,17 +3,15 @@ const express = require("express");
 const path = require("path");
 const compression = require("compression");
 const favicon = require("serve-favicon");
-const morgan = require("morgan");
 const session = require("express-session");
 const MongoStore = require("connect-mongo")(session);
 const bodyParser = require("body-parser");
 const credentials = require("./lib/credentials");
 const dbPromise = require("./lib/database").promise;
-const fs = require("fs-extra");
-const rfs = require("rotating-file-stream");
 const mongoSanitize = require("express-mongo-sanitize");
-const uuidv4 = require("uuid/v4");
-const bunyan = require("bunyan");
+
+const devEnv = require("./lib/devEnv");
+const { logger, applyLoggerMiddleware } = require("./lib/logger");
 
 //require route controllers
 const index = require("./routes/index");
@@ -24,19 +22,11 @@ const list = require("./routes/list");
 const sessionRoute = require("./routes/session");
 
 //make express app
-const app = express();
-
-//expose object with app router
-module.exports = {
-  router: app
-};
+const app = module.exports = express();
 
 //view engine setup
 app.set("views", path.join(__dirname, "views"));
 app.set("view engine", "pug");
-
-//check if we are in a dev environment, unefined if not set and in dev env
-const devEnv = ! process.env.NODE_ENV;
 
 //register express middleware
 app.use(compression()); //use compression to make it faster
@@ -46,84 +36,8 @@ app.use(mongoSanitize({
   replaceWith: "___" //something that can be easily found
 }));
 
-//complete console logging when in dev env
-if (devEnv) {
-  app.use(morgan("dev", {
-    //only log errors to console if in production
-    skip: (req, res) => devEnv || res.statusCode < 400
-  }));
-}
-
-//logging dir paths
-const loggingDir =  {
-  root: path.join(__dirname, "log")
-};
-loggingDir.access = path.join(loggingDir.root, "access");
-loggingDir.events = path.join(loggingDir.root, "events");
-
-//setup directory for logging
-try {
-  fs.ensureDirSync(loggingDir.access);
-} catch (err) {
-  console.error("failed to init logging directory");
-}
-
-//rotating logger config
-const logStreamConfig = {
-  interval: "1d", //rotate daily
-  size: "10M", //also rotate when the file reaches 10 megabytes
-  initialRotation: true //make sure log writes go in the right file
-};
-
-//rotating logger that makes a new file after a while
-const accessLogStream = rfs("access.log", Object.assign(
-  { path: loggingDir.access }, logStreamConfig
-));
-
-//give logger file stream to access logger middleware
-app.use(morgan("common", { stream: accessLogStream }));
-
-//events logger stream
-const eventsLogStream = rfs("access.log", Object.assign(
-  { path: loggingDir.events }, logStreamConfig
-));
-
-//array of logger streams
-const loggerStreams = [
-  {
-    //regular log to event file
-    level: "info",
-    stream: eventsLogStream
-  }
-];
-
-//add console output stream in dev mode
-if (devEnv) {
-  loggerStreams.push({
-    //dev stream to console
-    level: "debug",
-    stream: process.stdout
-  });
-}
-
-//setup bunyan event logger
-const logger = bunyan.createLogger({
-  name: "resolution-editor",
-  streams: loggerStreams,
-  serializers: bunyan.stdSerializers
-});
-
-//add logger to exported object
-module.exports.logger = logger;
-
-//add child logger as middleware to express
-app.use(function(req, res, next) {
-  //add bound field reqId to chain all log entries for a single request
-  req.log = logger.child({ req_id: uuidv4() });
-
-  //continue app stack
-  next();
-});
+//apply logger middlewares to app
+applyLoggerMiddleware(app);
 
 //static favicon serve
 app.use(favicon(path.join(__dirname, "public/favicon", "favicon.ico")));
