+++ lib/credentials.js
@@ -1,5 +1,6 @@
 /*jshint esversion: 6, node: true */
 const extend = require("extend");
+const { logger } = require("../lib/logger");
 
 //require keys file for set keys
 let keyFile;
@@ -7,7 +8,7 @@ try {
   keyFile = require("../lib/keys.json");
 } catch (err) {
   keyFile = {};
-  console.error("No lib/key.json file found, using default (and insecure) credentials!");
+  logger.warn("No lib/key.json file found, using default (and insecure) credentials!");
 }
 
 //extend cdefault credentials
