+++ lib/devEnv.js
@@ -1,0 +1,3 @@
+/*jshint esversion: 6, node: true */
+//check if we are in a dev environment, undefined if not set and in dev env
+module.exports = ! process.env.NODE_ENV;
